+++ src/v6/option_codes.rs
@@ -201,6 +201,7 @@ impl From<&DhcpOption> for OptionCode {
             O::InformationRefreshTime(_) => OptionCode::InformationRefreshTime,
             O::NtpServer(_) => OptionCode::NtpServer,
             O::ClientArchType(_) => OptionCode::ClientArchType,
+            O::ClientFqdn(_) => OptionCode::ClientFqdn,
             // SolMaxRt(_) => OptionCode::SolMaxRt,
             // InfMaxRt(_) => OptionCode::InfMaxRt,
             // LqQuery(_) => OptionCode::LqQuery,